requirements.txt
# API and Environment
requests>=2.31.0
python-dotenv>=1.0.0
betfairlightweight>=2.19.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Time handling
pytz>=2023.3

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Web UI
streamlit>=1.30.0

# Web Scraping (optional - for Selenium-based scripts)
selenium>=4.15.0
